Parallelized the travis build, migrated the generator to Python3.

+++ bindings/python/exporters_vec_types/__vec_exporter_10.cpp
@@ -15,6 +15,6 @@ using namespace fertilized;
 
 namespace pyfertilized {
   void __export_vectors_10() {
-    custom_vector_from_seq<std::shared_ptr<typename fertilized::Forest<float,float,float,std::pair<std::shared_ptr<std::vector<float>>,std::shared_ptr<std::vector<float>>>,std::vector<std::pair<std::pair<std::shared_ptr<std::vector<float>>,std::shared_ptr<std::vector<float>>>,float>>>::leaf_man_t>>();
+    custom_vector_from_seq<std::shared_ptr<typename fertilized::Forest<double,double,double,std::pair<std::shared_ptr<std::vector<double>>,std::shared_ptr<std::vector<double>>>,std::vector<std::pair<std::pair<std::shared_ptr<std::vector<double>>,std::shared_ptr<std::vector<double>>>,float>>>::leaf_man_t>>();
   };
 }  // namespace pyfertilized
