+++ bindings/python/exporters_vec_types/__vec_exporter_26.cpp
@@ -15,6 +15,6 @@ using namespace fertilized;
 
 namespace pyfertilized {
   void __export_vectors_26() {
-    custom_vector_from_seq<typename fertilized::Forest<float,float,uint,std::vector<float>,std::vector<float>>::tree_ptr_t>();
+    custom_vector_from_seq<typename fertilized::Forest<double,double,uint,std::vector<float>,std::vector<float>>::tree_ptr_t>();
   };
 }  // namespace pyfertilized
